readme – format bullet points
At "The shim consists of two parts:", the bullet points were previously formatted so that the `<pre>`s were rendered outside the `<li>`s.  This commit fixes that, which also lets the two rendered `<li>`s be in the same `<ul>`, which seems to more accurately represent the intended relationships.